fixup update rpm bug when using "-Uvh"

+++ rpm/sysak-build-nodep-aarch64.sh
@@ -57,7 +57,9 @@ if [ \$? -eq 0 ]; then
 fi
 
 %postun
-rm -rf /usr/local/sysak
+if [ \$1 -eq 0 ]; then
+	rm -rf /usr/local/sysak
+fi
 
 %files
 /etc/sysak
